PositionalLL.py
This linked list maintains the position of the elements, as object references. Which can be passed to replace or remove elements. 
However, i am not sure if i am using the functionality right.

This example is provided in the book DS and Algorithms in python by Goodrich, Tamassia and Goldwasser.